requirements.txt
Flask
requests
python-docx
PyMuPDF
serpapi
